requirements.txt
opencv-python==4.8.1.78
Pillow==10.0.1
PyQt5==5.15.10
numpy==1.24.3
rembg==2.0.50
